+++ h2integrate/converters/hydrogen/pem_electrolyzer.py
@@ -99,6 +99,8 @@ class ECOElectrolyzerPerformanceModel(ElectrolyzerPerformanceBaseClass):
         self.add_output("oxygen_out", val=0, shape=self.n_timesteps, units="kg/h")
         self.add_output("rated_oxygen_production", val=0, shape=1, units="kg/h")
         self.add_output("annual_oxygen_produced", val=0, shape=self.plant_life, units="kg/yr")
+        self.add_output("electricity_consumed", val=0, shape=self.n_timesteps, units="kW")
+        self.add_output("water_consumed", val=0, shape=self.n_timesteps, units="galUS/h")
 
         self.add_input("cluster_size", val=-1.0, units="MW")
         self.add_input("max_hydrogen_capacity", val=1000.0, units="kg/h")
@@ -169,6 +171,9 @@ class ECOElectrolyzerPerformanceModel(ElectrolyzerPerformanceBaseClass):
 
         # Assuming `h2_results` includes hydrogen and oxygen rates per timestep
         outputs["hydrogen_out"] = H2_Results["Hydrogen Hourly Production [kg/hr]"]
+        outputs["electricity_consumed"] = h2_ts.loc["Power Consumed [kWh]"]
+        # 1 gal H2O = 3.79 kg H2O
+        outputs["water_consumed"] = 3.79 * H2_Results["Water Hourly Consumption [kg/hr]"]
         outputs["total_hydrogen_produced"] = outputs["hydrogen_out"].sum()
         outputs["efficiency"] = H2_Results["Sim: Average Efficiency [%-HHV]"]
         refurb_schedule = np.zeros(self.plant_life)